app: improve compact spin-scale behavior
GimpSpinScale now always grabs focus in response to a button press
in compact mode.  However, like before, the left button is always
used for absolute adjustment, even when pressed over the text,
instead of controlling the text cursor.

The right button is now used both for relative adjustment, and to
select the entire text.  Alternatively, the same can be achieved
with Shift+left-button.

The middle button has no special function.

Regardless, if a button-press event over the text triggers a
context menu, this overrides any of the above.