app: draw border around FG/BG-editor color areas
In GimpFgBgEditor, we currently use gtk_render_frame() in master,
and gtk_paint_shadow() in gimp-2-10, to draw a border around the
color FG/BG color areas.  However, the former is relatively
subtle, especially with dark themes, and the latter is a NOP with
the pixmap engine, which is what our built-in themes use.

Instead, draw the border ourselves as a pair of black and white
rectangles, similarly to Photoshop.

Move the entire color-frame drawing functionality to a separate
function, to avoid code duplication between the FG and BG frames.