m4macros: update ax_gcc_func_attribute.m4 to serial 12
The serial 7 of m4macros/ax_gcc_func_attribute.m4 has an issue that when
Wstrict-prototypes is used with gcc, the attribute detection always
fails even if the attribute is actually supported. In such case the fish
cache will never be constructed. It's fixed in serial 12 of the macro in
autoconf-archive and this change updates the one distributed with GIMP
to serial 12.

(cherry picked from commit 7a4254dda1c0e514cfa0dbeb7f1b236688eb90fc)