app: do not lose multi item selection when dragging or right-clicking.
Basically the single click selection must happen on mouse button
release, not initial press, otherwise it would cancel your multiple
selection when you were actually about to drag the items.

As for contextual menu, it should trigger a selection only when the
clicked item is not in the current multiple selection.